feat: obcluster scale, update and upgrade
