+++ internal/cli/cli.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/oceanbase/ob-operator/internal/cli/cmd/cluster"
+	"github.com/oceanbase/ob-operator/internal/cli/cmd/install"
 	"github.com/oceanbase/ob-operator/internal/cli/cmd/tenant"
 	"github.com/oceanbase/ob-operator/internal/cli/cmd/version"
 	"github.com/spf13/cobra"
@@ -26,6 +27,7 @@ func NewCliCmd() *cobra.Command {
 	cmd.AddCommand(version.NewVersionCmd())
 	cmd.AddCommand(cluster.NewClusterCmd())
 	cmd.AddCommand(tenant.NewTenantCmd())
+	cmd.AddCommand(install.NewInstallCmd())
 	cmd.Flags().BoolP("version", "v", false, "Print the version number of oceanbase cli")
 	return cmd
 }
