+++ internal/cli/cmd/cluster/update.go
@@ -20,18 +20,20 @@ import (
 	cluster "github.com/oceanbase/ob-operator/internal/cli/pkg/cluster"
 	"github.com/oceanbase/ob-operator/internal/clients"
 	clusterstatus "github.com/oceanbase/ob-operator/internal/const/status/obcluster"
+	"github.com/oceanbase/ob-operator/internal/dashboard/business/constant"
 	oberr "github.com/oceanbase/ob-operator/pkg/errors"
 	"github.com/spf13/cobra"
+	apiresource "k8s.io/apimachinery/pkg/api/resource"
 )
 
-// NewUpgradeCmd upgrade obclusters
-func NewUpgradeCmd() *cobra.Command {
-	o := cluster.NewUpgradeOptions()
+// NewUpdateCmd update obcluster
+func NewUpdateCmd() *cobra.Command {
+	o := cluster.NewUpdateOptions()
 	logger := cmdUtil.GetDefaultLoggerInstance()
 	cmd := &cobra.Command{
-		Use:   "upgrade <cluster_name>",
-		Short: "Upgrade ob cluster",
-		Long:  "Upgrade ob cluster, please specify the new image",
+		Use:   "update <cluster_name>",
+		Short: "Update ob cluster",
+		Long:  "Update ob cluster, support update cpu/memory/storage",
 		Args:  cobra.ExactArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
 			o.Name = args[0]
@@ -45,7 +47,15 @@ func NewUpgradeCmd() *cobra.Command {
 			if obcluster.Status.Status != clusterstatus.Running {
 				logger.Fatalln(fmt.Errorf("Obcluster status invalid, Status:%s", obcluster.Status.Status))
 			}
-			obcluster.Spec.OBServerTemplate.Image = o.Image
+			if o.Cpu != 0 {
+				obcluster.Spec.OBServerTemplate.Resource.Cpu = *apiresource.NewQuantity(o.Cpu, apiresource.DecimalSI)
+			}
+			if o.MemoryGB != 0 {
+				obcluster.Spec.OBServerTemplate.Resource.Memory = *apiresource.NewQuantity(o.MemoryGB*constant.GB, apiresource.BinarySI)
+			}
+			if o.StorageGB != 0 {
+				obcluster.Spec.OBServerTemplate.Storage.DataStorage.Size = *apiresource.NewQuantity(o.StorageGB*constant.GB, apiresource.BinarySI)
+			}
 			cluster, err := clients.UpdateOBCluster(cmd.Context(), obcluster)
 			if err != nil {
 				logger.Fatalln(oberr.NewInternal(err.Error()))
