+++ internal/cli/cmd/util/helper.go
@@ -10,7 +10,7 @@ import (
 func PrintFlagValues(cmd *cobra.Command) {
 	// 确保命令的 flag 已经被解析
 	_ = cmd.ParseFlags(nil)
-	
+
 	// 遍历所有 flag
 	flags := cmd.NonInheritedFlags()
 	flags.VisitAll(func(f *pflag.Flag) {
