+++ internal/cli/pkg/cluster/generic.go
@@ -1,0 +1,6 @@
+package cluster
+
+type BaseOptions struct {
+	Name      string
+	Namespace string
+}
